types refactor
