+++ packages/core/src/types/callfabric.ts
@@ -12,190 +12,6 @@ import {
   VideoLayoutEvent,
 } from '..'
 
-export interface PaginatedResponse<T> {
-  data: Array<T>
-  links: {
-    first?: string
-    self?: string
-    next?: string
-    prev?: string
-  }
-}
-
-export interface PaginatedResult<T> {
-  data: Array<T> | []
-  self(): Promise<PaginatedResult<T> | undefined>
-  nextPage(): Promise<PaginatedResult<T> | undefined>
-  prevPage(): Promise<PaginatedResult<T> | undefined>
-  firstPage(): Promise<PaginatedResult<T> | undefined>
-  hasNext: boolean
-  hasPrev: boolean
-}
-
-/**
- * Addresses
- */
-
-export interface Address {
-  id: string
-  display_name: string
-  name: string
-  preview_url?: string
-  cover_url?: string
-  resource_id: string
-  type: string
-  channels: {
-    audio?: string
-    messaging?: string
-    video?: string
-  }
-}
-
-export interface GetAddressesParams {
-  type?: string
-  displayName?: string
-  pageSize?: number
-}
-
-export interface GetAddressParams {
-  id: string
-}
-
-export interface GetAddressResponse extends Address {}
-
-export interface FetchAddressResponse extends PaginatedResponse<Address> {}
-
-/**
- * Conversations
- */
-export interface SendConversationMessageParams {
-  text: string
-  addressId: string
-  metadata?: Record<string, any>
-  details?: Record<string, any>
-}
-
-export interface GetConversationsParams {
-  pageSize?: number
-}
-
-export interface Conversation {
-  created_at: number
-  id: string
-  last_message_at: number
-  metadata: Record<string, any>
-  name: string
-  sendMessage(params: {
-    text: string
-  }): Promise<SendConversationMessageResponse>
-  getMessages(params: {
-    pageSize?: number
-  }): Promise<PaginatedResult<ConversationMessage>>
-}
-
-export interface SendConversationMessageResponse {
-  table: {
-    conversation_id: string
-    text: string
-  }
-}
-
-export type SendConversationMessageResult =
-  Promise<SendConversationMessageResponse>
-
-export interface FetchConversationsResponse
-  extends PaginatedResponse<Conversation> {}
-
-/**
- * Conversation Messages
- */
-export interface GetMessagesParams {
-  pageSize?: number
-}
-
-export interface ConversationMessage {
-  id: string
-  conversation_id: string
-  user_id: string
-  ts: number
-  details: Record<string, any>
-  type: string
-  subtype: string
-  kind?: string
-  text?: string
-}
-
-export type GetMessagesResult = Promise<PaginatedResult<ConversationMessage>>
-
-export type ConversationChatMessage = Omit<ConversationMessage, 'kind'> & {
-  text: string
-}
-
-export interface GetConversationChatMessageParams {
-  addressId: string
-  pageSize?: number
-}
-
-export type GetConversationChatMessageResult = Promise<
-  PaginatedResult<ConversationChatMessage>
->
-
-export interface FetchConversationMessagesResponse
-  extends PaginatedResponse<ConversationMessage> {}
-
-export interface GetConversationMessagesParams {
-  addressId: string
-  pageSize?: number
-}
-
-export type GetConversationMessagesResult = Promise<
-  PaginatedResult<ConversationMessage>
->
-
-/**
- * Subsriber info
- */
-export interface SubscriberInfoResponse {
-  id: string
-  email: string
-  first_name?: string
-  last_name?: string
-  display_name?: string
-  job_title?: string
-  time_zone?: number
-  country?: string
-  region?: string
-  company_name?: string
-  push_notification_key: string
-  app_settings?: {
-    display_name: string
-    scopes: string[]
-  }
-}
-
-/**
- * Device registration
- */
-export type RegisterDeviceType = 'iOS' | 'Android' | 'Desktop'
-
-export interface RegisterDeviceParams {
-  deviceType: RegisterDeviceType
-  deviceToken: string
-}
-
-export interface UnregisterDeviceParams {
-  id: string
-}
-
-export interface RegisterDeviceResponse {
-  date_registered: Date
-  device_name?: string
-  device_token: string
-  device_type: RegisterDeviceType
-  id: string
-  push_notification_key: string
-}
-
 export type CallJoined = 'call.joined'
 export type CallStarted = 'call.started'
 export type CallUpdated = 'call.updated'
